Refactor: Extract score normalization to separate file
- Move score normalization functions to score.go
- Move score normalization tests to score_test.go
- Add detailed comments explaining intentional code duplication
- Remove duplicate code from anime.go and manga.go
- Remove duplicate tests from anime_test.go and manga_test.go

Addresses Copilot review comments about documenting intentional
duplication in fetch functions.
