+++ .gitignore
@@ -1,8 +1,13 @@
+bin/
+anilist-mal-sync
+
+.env
 .token
 .token*
-bin/
-config.yaml
-*tmp.json
-CLAUDE.md
 .claude
-anilist-mal-sync
+
+config.yaml
+CLAUDE.md
+
+*tmp.json
+*.log
