+++ anime.go
@@ -534,91 +534,6 @@ func newAnimeFromVerniyMedia(media verniy.Media) (Anime, error) {
 	}, nil
 }
 
-// normalizeScoreForMAL converts AniList score to 0-10 int format
-func normalizeScoreForMAL(anilistScore float64, scoreFormat verniy.ScoreFormat) int {
-	if anilistScore == 0 {
-		return 0
-	}
-
-	var normalized float64
-
-	switch scoreFormat {
-	case verniy.ScoreFormatPoint100:
-		normalized = anilistScore / 10.0
-		DPrintf("Score normalized (POINT_100): %.1f → %.1f", anilistScore, normalized)
-
-	case verniy.ScoreFormatPoint100Decimal: // "POINT_10_DECIMAL"
-		normalized = anilistScore
-
-	case verniy.ScoreFormatPoint10:
-		normalized = anilistScore
-
-	case verniy.ScoreFormatPoint5:
-		normalized = anilistScore * 2.0
-		DPrintf("Score normalized (POINT_5): %.1f → %.1f", anilistScore, normalized)
-
-	case verniy.ScoreFormatPoint3:
-		normalized = (anilistScore / 3.0) * 10.0
-		DPrintf("Score normalized (POINT_3): %.1f → %.1f", anilistScore, normalized)
-
-	default:
-		if anilistScore > 10 {
-			normalized = anilistScore / 10.0
-			DPrintf("Score normalized (unknown): %.1f → %.1f", anilistScore, normalized)
-		} else {
-			normalized = anilistScore
-		}
-	}
-
-	// Clamp 0-10
-	if normalized > 10 {
-		DPrintf("Score %.1f > 10, clamping", normalized)
-		normalized = 10
-	}
-	if normalized < 0 {
-		normalized = 0
-	}
-
-	return int(normalized + 0.5) // Round to nearest int
-}
-
-// denormalizeScoreForAniList converts 0-10 int format back to AniList format
-func denormalizeScoreForAniList(normalizedScore int, scoreFormat verniy.ScoreFormat) int {
-	if normalizedScore == 0 {
-		return 0
-	}
-
-	var result float64
-
-	switch scoreFormat {
-	case verniy.ScoreFormatPoint100:
-		result = float64(normalizedScore) * 10.0
-		DPrintf("Score denormalized (POINT_100): %d → %.1f", normalizedScore, result)
-
-	case verniy.ScoreFormatPoint100Decimal: // "POINT_10_DECIMAL"
-		result = float64(normalizedScore)
-
-	case verniy.ScoreFormatPoint10:
-		result = float64(normalizedScore)
-
-	case verniy.ScoreFormatPoint5:
-		result = float64(normalizedScore) / 2.0
-		DPrintf("Score denormalized (POINT_5): %d → %.1f", normalizedScore, result)
-
-	case verniy.ScoreFormatPoint3:
-		result = (float64(normalizedScore) / 10.0) * 3.0
-		if result < 1 && normalizedScore > 0 {
-			result = 1
-		}
-		DPrintf("Score denormalized (POINT_3): %d → %.1f", normalizedScore, result)
-
-	default:
-		result = float64(normalizedScore)
-	}
-
-	return int(result + 0.5) // Round to nearest int
-}
-
 func buildDiffString(pairs ...any) string {
 	if len(pairs)%3 != 0 {
 		return "Diff{invalid params}"
